+++ iommufd-ioctls/CHANGELOG.md
@@ -5,6 +5,8 @@
 ## Added
 - [[9]](https://github.com/cloud-hypervisor/iommufd/pull/9) Add `Iommufd::destroy_iommu_object` to release iommufd
   objects explicitly.
+- [[9]](https://github.com/cloud-hypervisor/iommufd/pull/9) Add `Iommufd::new_from_fd` to construct a `Iommufd` from
+  a pre-opened iommufd file.
 
 ## Fixed